Added some help info for NuGet use.
I accidentally made the mistake of trying to use NuGet in the solution folder and it took a few minutes to figure out what I was doing wrong.

So, I added this sentence here to help clarify what to do for this (probably common) problem since Visual Studio is fairly likely to start the working directory for the command prompt in the solution directory, which may not be the same as the project directory. 

I think since raylib attracts many new users it is best to make the install info as easy as possible, thus my justification for adding this here. 

Installs should be as frictionless as possible. I hope you'll consider adding my suggestion here. 🙂